apps/ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
scikit-learn==1.6.0
numpy==2.1.0
pandas==2.2.0
httpx==0.28.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-exporter-otlp==1.28.0
python-multipart==0.0.12
